Move MarkdownPage class into Models\Pages namespace
