+++ packages/framework/src/Services/CollectionService.php
@@ -6,7 +6,7 @@ use Hyde\Framework\Contracts\AbstractPage;
 use Hyde\Framework\Hyde;
 use Hyde\Framework\Models\BladePage;
 use Hyde\Framework\Models\DocumentationPage;
-use Hyde\Framework\Models\MarkdownPage;
+use Hyde\Framework\Models\Pages\MarkdownPage;
 use Hyde\Framework\Models\MarkdownPost;
 
 /**
