+++ packages/framework/src/StaticPageBuilder.php
@@ -7,7 +7,7 @@ use Hyde\Framework\Concerns\InteractsWithDirectories;
 use Hyde\Framework\Contracts\PageContract;
 use Hyde\Framework\Models\BladePage;
 use Hyde\Framework\Models\DocumentationPage;
-use Hyde\Framework\Models\MarkdownPage;
+use Hyde\Framework\Models\Pages\MarkdownPage;
 use Hyde\Framework\Models\MarkdownPost;
 
 /**
