+++ packages/framework/tests/Feature/Services/CollectionServiceTest.php
@@ -4,7 +4,7 @@ namespace Hyde\Framework\Testing\Feature\Services;
 
 use Hyde\Framework\Hyde;
 use Hyde\Framework\Models\DocumentationPage;
-use Hyde\Framework\Models\MarkdownPage;
+use Hyde\Framework\Models\Pages\MarkdownPage;
 use Hyde\Framework\Models\MarkdownPost;
 use Hyde\Framework\Services\CollectionService;
 use Hyde\Testing\TestCase;
