+++ packages/framework/tests/Unit/PageModelParseHelperTest.php
@@ -5,7 +5,7 @@ namespace Hyde\Framework\Testing\Unit;
 use Hyde\Framework\Hyde;
 use Hyde\Framework\Models\BladePage;
 use Hyde\Framework\Models\DocumentationPage;
-use Hyde\Framework\Models\MarkdownPage;
+use Hyde\Framework\Models\Pages\MarkdownPage;
 use Hyde\Framework\Models\MarkdownPost;
 use Hyde\Testing\TestCase;
